watermarker/requirements.txt
boto3>=1.34.0
botocore>=1.34.0
Pillow>=10.0.0
steganography>=0.1.1
requests>=2.31.0